# docs/openapi/flagman-delivery.openapi.yaml
openapi: 3.0.3
info:
  title: Flagman Delivery API
  version: 1.0.0
  description: |
    OpenAPI specification generated from the FastAPI backend in `hac.new` and normalized for iOS client generation.
    The backend exposes JWT bearer authentication via `/api/auth/login` and `/api/auth/refresh`.
    Password reset and email verification flows are also described so the spec can be used for iOS client generation.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/couriers/profile:
    get:
      tags: [couriers]
      operationId: getCourierProfile
      summary: Get current courier profile
      responses:
        '200':
          description: Courier profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourierProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [couriers]
      operationId: upsertCourierProfile
      summary: Update courier profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourierProfileUpdate'
            example:
              phone: "+79994443322"
              vehicle_type: bicycle
              license_plate: null
      responses:
        '200':
          description: Updated courier profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourierProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/couriers/shift:
    patch:
      tags: [couriers]
      operationId: toggleCourierShift
      summary: Toggle courier online/offline shift status
      responses:
        '200':
          description: Shift state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourierShiftResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/couriers/available-orders:
    get:
      tags: [couriers]
      operationId: listCourierAvailableOrders
      summary: List available orders for online courier
      responses:
        '200':
          description: Available orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/couriers/orders/{order_id}/accept:
    post:
      tags: [couriers]
      operationId: acceptCourierOrder
      summary: Accept available ready order
      parameters:
        - $ref: '#/components/parameters/OrderID'
      responses:
        '200':
          description: Accepted order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/couriers/current-order:
    get:
      tags: [couriers]
      operationId: getCourierCurrentOrder
      summary: Get currently assigned accepted order
      responses:
        '200':
          description: Current order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/couriers/current-order/status:
    patch:
      tags: [couriers]
      operationId: updateCourierCurrentOrderStatus
      summary: Mark current order as picked up or delivered
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourierCurrentOrderStatusRequest'
            example:
              status: delivered
      responses:
        '200':
          description: Updated current order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/couriers/history:
    get:
      tags: [couriers]
      operationId: getCourierHistory
      summary: Get delivered history for current courier
      responses:
        '200':
          description: Courier history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourierHistoryOrderItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/couriers/location:
    post:
      tags: [couriers]
      operationId: updateCourierLocation
      summary: Send current courier GPS position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdate'
            example:
              lat: 55.7601
              lon: 37.6112
      responses:
        '200':
          description: Updated courier profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourierProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderID:
      name: order_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Адрес доставки вне зоны доставки магазина
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Недействительный или просроченный токен
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Требуется роль бизнес-пользователя
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Заказ не найден
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Заказ недоступен или уже назначен
...
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
    VehicleType:
      type: string
      enum: [foot, bicycle, motorcycle, car]
    OrderStatus:
      type: string
      enum: [draft, pending, confirmed, preparing, ready, assigned, picked_up, on_the_way, delivered, cancelled]
    CourierAvailability:
      type: string
      enum: [offline, online, busy]
...
    CourierDeliveryStatusAction:
      type: string
      enum: [picked_up, delivered]
    Coordinate:
      type: object
      required: [lat, lon]
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
...
    CourierProfileUpdate:
      type: object
      properties:
        phone:
          type: string
          nullable: true
        vehicle_type:
          $ref: '#/components/schemas/VehicleType'
        license_plate:
          type: string
          nullable: true
    CourierProfileResponse:
      allOf:
        - $ref: '#/components/schemas/CourierProfileUpdate'
        - type: object
          required: [id, user_id, availability]
          properties:
            id:
              type: integer
            user_id:
              type: integer
            availability:
              $ref: '#/components/schemas/CourierAvailability'
            current_lat:
              type: number
              format: double
              nullable: true
            current_lon:
              type: number
              format: double
              nullable: true
...
    OrderLineSnapshot:
      type: object
      properties:
        item_id:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        unit_price:
          type: string
          format: decimal
        line_total:
          type: string
          format: decimal
    ItemsSnapshot:
      type: object
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineSnapshot'
...
    CourierShiftResponse:
      type: object
      required: [availability]
      properties:
        availability:
          $ref: '#/components/schemas/CourierAvailability'
    AvailableOrderResponse:
      type: object
      required: [id, store_name, distance_km, reward]
      properties:
        id:
          type: integer
        store_name:
          type: string
        store_address:
          type: string
          nullable: true
        delivery_address:
          type: string
          nullable: true
        distance_km:
          type: number
          format: double
        reward:
          type: string
          format: decimal
    CourierOrderCustomerContact:
      type: object
      properties:
        full_name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
    AcceptOrderResponse:
      type: object
      required: [id, public_id, status, subtotal, delivery_fee, total, created_at, updated_at, customer, store_name]
      properties:
        id:
          type: integer
        public_id:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        items_snapshot:
          $ref: '#/components/schemas/ItemsSnapshot'
        delivery_address:
          type: string
          nullable: true
        delivery_coordinates:
          $ref: '#/components/schemas/Coordinate'
        comment:
          type: string
          nullable: true
        subtotal:
          type: string
          format: decimal
        delivery_fee:
          type: string
          format: decimal
        total:
          type: string
          format: decimal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        customer:
          $ref: '#/components/schemas/CourierOrderCustomerContact'
        store_name:
          type: string
        store_address:
          type: string
          nullable: true
        store_phone:
          type: string
          nullable: true
    CourierCurrentOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/CourierDeliveryStatusAction'
    CourierHistoryOrderItem:
      type: object
      required: [id, public_id, status, total, delivery_fee, created_at, updated_at, store_name]
      properties:
        id:
          type: integer
        public_id:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        total:
          type: string
          format: decimal
        delivery_address:
          type: string
          nullable: true
        delivery_fee:
          type: string
          format: decimal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        store_name:
          type: string
    LocationUpdate:
      type: object
      required: [lat, lon]
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
...